+++ Plugins/Gradle/src/test/kotlin/dev/testify/internal/AdbTest.kt
@@ -25,7 +25,7 @@ package dev.testify.internal
 
 import com.android.build.gradle.TestedExtension
 import com.google.common.truth.Truth.assertThat
-import io.mockk.MockKAnnotations
+import dev.testify.test.BaseTest
 import io.mockk.every
 import io.mockk.impl.annotations.RelaxedMockK
 import io.mockk.mockkObject
@@ -40,7 +40,7 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import java.io.File
 
-class AdbTest {
+class AdbTest : BaseTest() {
 
     @RelaxedMockK
     lateinit var project: Project
@@ -61,11 +61,12 @@ class AdbTest {
     private lateinit var subject: Adb
 
     @BeforeEach
-    fun setUp() {
-        MockKAnnotations.init(this)
+    override fun setUp() {
+        super.setUp()
 
         every { extension.adbExecutable } returns adbExecutable
 
+        mockkStatic("dev.testify.internal.ClientUtilitiesKt")
         mockkStatic(Project::android)
         mockkStatic(Project::isVerbose)
         mockkStatic(Project::user)
@@ -76,6 +77,9 @@ class AdbTest {
         every { any<Project>().user } returns null
 
         mockkStatic(::runProcess)
+        mockkObject(::isEnvSet)
+
+        every { isEnvSet(any()) } returns true
 
         configureRunProcessCapture(defaultResultMap)
         Adb.init(project)